# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于 Gin + GORM + Redis + RabbitMQ 的在线书店后端服务，提供用户认证、书籍管理、购物车、订单及秒杀等功能。",
        "title": "Project1_Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Me",
            "email": "frottenice@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
        "/admin/book/add": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员添加新书籍（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍管理（管理员）"
                ],
                "summary": "添加书籍（管理员）",
                "parameters": [
                    {
                        "description": "书籍信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddBookParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 无权限 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/admin/book/delete/{book_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员根据书籍 ID 删除书籍（需要管理员权限）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍管理（管理员）"
                ],
                "summary": "删除书籍（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "书籍 ID",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 书籍不存在 / 无权限 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/admin/book/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员更新书籍信息，book_id 为必填（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍管理（管理员）"
                ],
                "summary": "更新书籍信息（管理员）",
                "parameters": [
                    {
                        "description": "更新书籍参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateBookParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 书籍不存在 / 无权限 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/comment": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户对书籍发表评论，支持嵌套回复（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "发表书籍评论",
                "parameters": [
                    {
                        "description": "评论参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CommentParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/comment/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户对指定评论进行点赞（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "点赞评论",
                "parameters": [
                    {
                        "description": "评论点赞参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CommentLikeParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/comments": {
            "get": {
                "description": "根据书籍 ID 获取该书籍的所有评论（树形结构）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "获取书籍评论列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "书籍 ID",
                        "name": "book_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/getBookDetail/{book_id}": {
            "get": {
                "description": "通过书籍 ID 获取书籍详细信息（管理员和普通用户均可访问）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "根据书籍 ID 获取详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "书籍 ID",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 书籍不存在 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/getBookTitle/{title}": {
            "get": {
                "description": "通过书名关键字模糊搜索书籍列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "根据书名搜索书籍",
                "parameters": [
                    {
                        "type": "string",
                        "description": "书名关键字",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "书籍不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/getBooksBySaleJSON/{page}": {
            "get": {
                "description": "按书籍销量降序分页返回书籍列表，指定页码（每页 8 条）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "按销量分页获取书籍（指定页码）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（从 1 开始）",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/getBooksJSON/{page}": {
            "get": {
                "description": "按书籍评分降序分页返回书籍列表，指定页码（每页 8 条）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "按评分分页获取书籍（指定页码）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码（从 1 开始）",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/rateBook": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户对书籍进行评分（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "书籍"
                ],
                "summary": "对书籍评分",
                "parameters": [
                    {
                        "description": "评分参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserRateBook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddBookParam": {
            "type": "object",
            "required": [
                "author",
                "price",
                "publisher",
                "stock",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "cover_image": {
                    "type": "string"
                },
                "introduction": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "publisher": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "models.CommentLikeParam": {
            "type": "object",
            "required": [
                "comment_id"
            ],
            "properties": {
                "comment_id": {
                    "type": "integer"
                }
            }
        },
        "models.CommentParam": {
            "type": "object",
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "comment": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "root_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {}
            }
        },
...
        "models.UpdateBookParam": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "book_id": {
                    "type": "integer"
                },
                "cover_image": {
                    "type": "string"
                },
                "introduction": {
                    "type": "string"
                },
                "price": {
                    "description": "Price和Score都用int64，前端显示时除以100,避免精度误差",
                    "type": "integer"
                },
                "publisher": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "tags": {
                    "description": "切片用法",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.UserRateBook": {
            "type": "object",
            "properties": {
                "Op": {
                    "type": "integer"
                },
                "book_id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}